modules/ike-processor/requirements.txt
pdfplumber==0.7.6
pandas==1.5.3
openpyxl==3.1.2
pymongo==4.3.3
python-dotenv==1.0.0
